+++ src/agent/common/haylee-tool.base.ts
