+++ src/agent/subagents/tools/README.md
@@ -1,0 +1,5 @@
+This subfolder defines tools that are SHARED across subagents.
+
+All tools should still follow Nest DI pattern so we stay inside the DI container.
+
+Thanks
