--- src/agent/tools/factory/index.ts
@@ -1,1 +1,0 @@
-export * from "./haylee-tool.factory";
