+++ src/agent/tools/implementation/launch-researcher-subagent.impl.ts
@@ -1,5 +1,5 @@
 import { StructuredTool, tool } from "langchain";
-import { HayleeTool } from "../base";
+import { HayleeTool } from "../../common";
 import z from "zod";
 import { Inject, Injectable } from "@nestjs/common";
 import { IRunnableSubagent, SubagentMap } from "src/agent/subagents/interface";
