+++ src/agent/tools/implementation/upload_medialibrary_to_meta.impl.ts
@@ -1,5 +1,5 @@
 import { StructuredTool, tool } from "langchain";
-import { HayleeTool } from "../base";
+import { HayleeTool } from "../../common";
 import { RunnableConfig } from "@langchain/core/runnables";
 import z from "zod";
 import { QueryAdCreativeDtoSchema } from "src/generated/schemas/create-adcreative.dto.schema";
